requirements.txt
about-time==4.2.1
aiofiles==23.1.0
alive-progress==3.1.5
anyio==3.7.1
argcomplete==3.1.1
attrs==23.1.0
autograd==1.6.2
boto3==1.28.15
botocore==1.31.15
Brotli==1.0.9
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
cma==3.2.2
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
dacite==1.7.0
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.1
eventlet==0.33.3
exceptiongroup==1.1.2
fiftyone==0.21.4
fiftyone-brain==0.13.0
fiftyone-db==0.4.0
filelock==3.9.0
fonttools==4.41.1
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
glob2==0.7
grapheme==0.6.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
importlib-resources==6.0.0
inflate64==0.3.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonlines==3.1.0
kaleido==0.2.1
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.2
mongoengine==0.24.2
motor==3.2.0
mpmath==1.2.1
multivolumefile==0.2.3
networkx==3.0
numpy==1.24.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
plotly==5.15.0
pprintpp==0.4.0
priority==2.0.0
psutil==5.9.5
py7zr==0.20.5
pybcj==1.0.1
pycryptodomex==3.18.0
pymongo==4.4.1
pymoo==0.6.1.1
pyparsing==3.0.9
pyppmd==1.0.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzstd==0.15.9
rarfile==4.0
regex==2023.6.3
requests==2.28.1
retrying==1.3.4
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.26.1
strawberry-graphql==0.138.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.18
timm==0.9.2
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.0.1
universal-analytics-python3==1.1.1
urllib3==1.26.13
voxel51-eta==0.10.0
wcwidth==0.2.6
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.16.2
